library_management_system/backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4  # Or PyMySQL==1.1.1 if mysqlclient fails
scikit-learn==1.5.0  # For AI recommendations
numpy==2.0.0
pandas==2.2.2
django-cors-headers==4.4.0  # For React CORS
Pillow==10.4.0  # For any image handling (optional)
joblib==1.3.2  # For model serialization